requirements-test.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# HTTP & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
curl_cffi>=0.6.0  # JavDB compatibility

# WebSocket
websockets>=12.0

# Utilities
pillow>=10.0.0

# Testing & Dev
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
mypy>=1.0.0
types-requests
pytest-playwright>=0.4.0  # e2e testing
